move login_required to source_app/decorators.py
