+++ securedrop/source_app/decorators.py
@@ -1,0 +1,13 @@
+from flask import redirect, url_for
+from functools import wraps
+
+from source_app.utils import logged_in
+
+
+def login_required(f):
+    @wraps(f)
+    def decorated_function(*args, **kwargs):
+        if not logged_in():
+            return redirect(url_for('login'))
+        return f(*args, **kwargs)
+    return decorated_function
